+++ packages/form_bloc/test/boolean_field/boolean_field_bloc_test.dart
@@ -150,7 +150,7 @@ void main() {
         );
         final state2 = state1.copyWith(
           value: Optional.of(false),
-          isInitial: false,
+          isInitial: true,
         );
 
         final expectedStates = [