+++ packages/form_bloc/test/text_field/text_field_bloc_test.dart
@@ -125,7 +125,7 @@ void main() {
       );
       final state2 = state1.copyWith(
         value: Optional.of(''),
-        isInitial: false,
+        isInitial: true,
       );
 
       final expectedStates = [